Restructure files and libraries
	Rename a bunch of files to make the consistent
	Create two new subdirectories for common code, and utilities
