+++ common/common-external.c
@@ -31,8 +31,8 @@ POSSIBILITY OF SUCH DAMAGE.
 
 #define _GNU_SOURCE
 
-#include "external-helper.h"
-#include "logging.h"
+#include "common-external.h"
+#include "common-logging.h"
 
 #include <linux/limits.h>
 #include <stdio.h>
