+++ common/common-external.h
