+++ common/common-helpers.h
