+++ common/common-logging.h
