+++ daemon/gamemode-config.c
@@ -30,9 +30,10 @@ POSSIBILITY OF SUCH DAMAGE.
  */
 #define _GNU_SOURCE
 
-#include "daemon_config.h"
-#include "helpers.h"
-#include "logging.h"
+#include "gamemode-config.h"
+
+#include "common-helpers.h"
+#include "common-logging.h"
 
 /* Ben Hoyt's inih library */
 #include "ini.h"
