+++ daemon/gamemode-config.h
