+++ daemon/gamemode-gpu.c
@@ -32,15 +32,14 @@ POSSIBILITY OF SUCH DAMAGE.
 
 #define _GNU_SOURCE
 
-#include "config.h"
-#include "external-helper.h"
-#include "helpers.h"
-#include "logging.h"
-
+#include "common-external.h"
+#include "common-gpu.h"
+#include "common-helpers.h"
+#include "common-logging.h"
+#include "gamemode-config.h"
 #include "gamemode.h"
 
-#include "daemon_config.h"
-#include "gpu-control.h"
+#include "build-config.h"
 
 /**
  * Attempts to identify the current in use GPU information
